Cleanup relation form
Relation toolbar was never used in the category form, so this updates
the field definition for categories so we can remove the toolbar partial.

While we're at it, let's get rid of the Manage tab, too. It's the only
tab on the form, so there's no sense in having it there.