Fix product creation redirect bug
Product create form wasn't redirecting to update, which in turn
caused a duplicate key SQL error that broke things. After 3 days
of not seeing this line despite looking at the file, it's finally
fixed by changing a single bloody digit. I'm an idiot.

We don't want the redirect on the update form though, so let's
simply force $isCreate to an integer so that it only redirects
if the user is on the create form. Using the plain bool works
for create, but doesn't show the 0 on update.